net/txgbe: fix FDIR devarg integer width
txgbe_handle_devarg() stores a uint16_t through extra_args, but
pballoc and drop_queue were u8. Setting pkt-filter-size overwrote
the adjacent drop_queue default.

Fixes: 7e18be9beef2 ("net/txgbe: add device arguments for FDIR")
Cc: stable@dpdk.org

+++ drivers/net/txgbe/txgbe_ethdev.c
@@ -479,8 +479,8 @@ txgbe_parse_devargs(struct rte_eth_dev *dev)
 	u16 ffe_pre = 8;
 	u16 ffe_post = 44;
 	/* FDIR args */
-	u8 pballoc = 0;
-	u8 drop_queue = 127;
+	u16 pballoc = 0;
+	u16 drop_queue = 127;
 	/* New devargs for amberlite config */
 	u16 tx_headwb = 1;
 	u16 tx_headwb_size = 16;
